Move ExtractFrozenUpdate into lib/Update.py, and add an exception for it.  This should make it easier for the GUI (or other clients) to extract an update.
Ticket # 29365
